requirements.txt
# Flask Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10
gunicorn>=21.2.0

# Core ML/AI
langchain>=0.1.0
langchain-community>=0.0.20

# Ollama Integration
ollama>=0.1.0

# Vector Database
chromadb>=0.4.0

# Document Processing
pdfplumber>=0.10.0
pypdf2>=3.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Evaluation
ragas>=0.1.0
scikit-learn>=1.3.0

# Text Processing
nltk>=3.8.0

# API Documentation
flasgger>=0.9.7

# Validation
marshmallow>=3.20.0

# Caching
redis>=5.0.0
